refactor(runner): decompose lab_workspaces god-file into submodules
lab_workspaces.rs was 2903 lines (god_file). Unlike the prior
decompositions, its only inline seams were three #[cfg(test)] modules;
the 1595-line production region had no pre-drawn seams, so extracting the
tests alone left mod.rs still over threshold.

Two extractions:
- Move the three test modules into lab_workspaces/tests/{provider_config,
  runtime_overlay,validation_dependency}.rs (the first was ~1080 lines on
  its own), gathered under tests/mod.rs. super:: paths adjusted for the
  added nesting level.
- Extract the cohesive workspace-ref path-rewriting cluster
  (push_path_setting_value, rewrite_path_setting_*, rewrite_workspace_*,
  maybe_resolve_workspace_ref, parse_workspace_ref, subcommand_index)
  into workspace_ref_rewrite.rs, exposed to the parent via pub(super) +
  use workspace_ref_rewrite::*.

mod.rs is now 1373 lines; every resulting file is under the 1500-line
god_file threshold. No logic changed; public API preserved.

Clears the lab_workspaces.rs god_file finding.
