+++ src/core/runner/lab_workspaces/tests/mod.rs
@@ -1,0 +1,3 @@
+mod provider_config;
+mod runtime_overlay;
+mod validation_dependency;
